Header element - Creder header element - Import from header element - Remove from index
